Move account closing dialog to a new package
