+++ WooCommerce/src/main/res/navigation/nav_graph_site_picker.xml
@@ -89,6 +89,6 @@
     </dialog>
     <dialog
         android:id="@+id/closeAccountDialogFragment"
-        android:name="com.woocommerce.android.ui.prefs.CloseAccountDialogFragment"
+        android:name="com.woocommerce.android.ui.prefs.account.CloseAccountDialogFragment"
         android:label="CloseAccountDialogFragment" />
 </navigation>
